# flask_api/swagger.yml
...
openapi: 3.0.0
info:
  title: "Scientific Flask REST API"
  description: "An API for interacting with the scientific database"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
    Sci:
      type: "object"
      properties:
        id:
          type: "integer"
          nullable: true  # optional in POST requests.
        date:
          type: "string"
        time:
          type: "string"
        time_offset:
          type: "integer"
        coordinate: 
          type: "array"
          items:
            type: "number"
        air_temperature:
          type: "number"
        humidity:
          type: "number"
        wind_speed:
          type: "number"
        wind_direction:
          type: "number"
        precipitation:
          type: "number"
        haze:
          type: "number"
        water_temperature:
          type: "number"
        notes:
          type: "string"
  parameters:
    id:
      name: "id"
      description: "Unique id of the record to get"
      in: path
      required: True
      schema:
        type: "integer"
paths:
...
  /sci_record/{id}:
    get:
      operationId: "sci_record.read_one"
      tags:
        - Sci
      summary: "Read one record"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "Successfully read record with id"
    put:
      tags:
        - Sci
      operationId: "sci_record.update"
      summary: "Update a record"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "Successfully updated record"
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: "sci_record"
              $ref: "#/components/schemas/Sci"
    delete:
      tags:
        - Sci
      operationId: "sci_record.delete"
      summary: "Delete a record"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: "Successfully deleted record"